cdk/requirements.txt
aws-cdk-lib==2.87.0
constructs>=10.0.0,<11.0.0
pepperize.cdk-organizations==0.7.577
